Add tests for Game and WordleGame.__init__.py
Using pytest and pytest-mock for unit tests since they're familiar and
still seem to be among the best options.

Some of these currently tests fail due to bugs or poorly defined
behaviour identified by them. Some messing with PYTHONPATH is required
to run the tests until the ability to install an editable package is set
up.

As always, trying to find the right balance between avoiding boilerplate
noise in tests and avoiding making them too complex and hard to read -
in my experience the best way to have tests that work correctly is to
ensure that what they actually do is obvious. The mock_word_list fixture
is pretty borderline in this regard, but it does make the test cases
nice and simple.

Testing it makes it more obvious that the way I've implemented ALPHABET
as an abstract method is a bit weird, but it's the best way I've thought
of so far that actually ensures it gets implemented by a subclass. The
tests also make me think that the word list should probably be injected
somehow, maybe I'll revisit that later.