Add venv creation using pyenv and pip-tools
Pyenv allows using different python versions (to avoid messing with
system version or for testing other versions) and pip-tools allows
compiling fixed requirements from pyproject.toml dependencies. Simple
custom scripts make using them easier.

I'm not certain if pip-compile will be especially meaningful for a
library (as opposed to application) where I want to ensure the package
installs and runs with latest dependency versions, but it might be
useful for reproducable dev envs or testing in multiple environments
(e.g. last known working + latest); and I want to try it out anyway.
Similarly, I'm not quite sure if including .python-version in the repo
will be meaningful in this case. Both will be clearer once I have CI set
up.

Next step is to make the package buildable using pyproject.toml and
and a build backend, then work out how to include an editable install in
this venv setup.

Ultimately I'd like to try poetry instead of some of these tools, and,
if I like it, switch to it. But I decided to do it this way first to see
both approaches.