Use editable package install in dev env
I'm starting to get a clearer idea of where the editable install might
actually be useful, compared to just installing the package and getting
the latest dependencies, so I've written about it in the README.

This also allows the tests to be run simply without changing PYTHONPATH,
using the installed editable package. So we can test changes in
progress, but also ensure that the tests match the user installation.