Add pre-commit
The way that pre-commit installs tools via git tags and ignores the
environment in use is neat but kind of annoying, since it means either
duplicating them, only using the tool via pre-commit, or bypassing the
typical setup with a "system" pre-commit configuration. For ruff I've
removed it from the venv and can only use it via pre-commit, which seems
fine since I only use it in simple ways, but could become annoying
later.

In the case of mypy I immediately ran into a problem with the fact that
I would need to specify my dependencies (or type stubs for them) in the
pre-commit configuration to get it to work correctly in the isolated
environment. I absolutely don't want that duplication, so instead I'm
using the "system" workaround, which isn't much extra configuration.
This follows the logic described here:
https://github.com/python/mypy/issues/13916. I also disabled
pass_filenames due to it not making much sense with mypy, as described
there.

Another pain point is needing to have pre-commit available to commit,
which right now means activating the appropriate environment. Probably a
case where pipx would be ideal, but I don't feel like setting it up
right now.