Change word list to "word_dictionary" set
Checking valid words with each guess makes it clear that this should be
a set. There's now a slightly inefficient conversion to list for the
target random selection, but the act of checking whether a word is valid
will be much quicker regardless of dictionary size. The new name also
fits the purpose a bit better (even if it clashes a bit with Python
dicts).