Implement optional specific word length
This mainly works by just filtering the dictionary down to words of that
length. Only guesses of that length will be allowed because only words
of that length will be in the game dictionary.

I thought this would be a nice place to use the new(ish) walrus operator
to strip() words and filter by length in the same comprehension, but the
order or operations and the fact that the condition has two parts made
it not work. I could've swapped the condition order and had it work, but
relying on the order like that felt non-obvious and likely to be
confusing.

Changed the mock dictionary file fixtures back to returning lists rather
than sets to more easily make it clear that the dictionary file contents
may not be the same as the game dictionary.