Impement guess letter status calculation
In wordle each letter in a guess can be either "correct", "incorrect",
or "correct but in the wrong place". The logic for this has to account
for duplicate letters. In my case the logic also needs to account for
the guess and target being different lengths, since the WordleGame
doesn't enforce length in the default case - luckily this is quite
simple.

It became obvious quite quickly (when writing type hints) that having
the list of guesses use basic data structures was going to have too much
nesting, so I separated out the concept of a Guess into a new class.
This works quite nicely since I can put the status calculation logic
there as well.

It's not clear to me yet how other game implementations will handle
guesses - will they need a different type of guess? They may well need
to have multiple Guess objects per guess (e.g. quordle). How will the
CLI interact with the guesses for different games? This might need
refactoring a bunch later.