Add alphabet letter states to wordle game
In wordle the state of each letter in the alphabet should be visible to
guide guesses. This is related to the state of each letter in the guess,
but not the same.

The tests for this feature really required explicitly setting the target
and guesses. I liked how it looked better than picking out of the
original word list, so I changed the other tests to match.