Only allow guesses with same length as target
Even if a specific target length isn't chosen, Wordle now only allows
guesses with the same length as the target. This makes sense in terms of
gameplay - it would be much easier if you could guess a long word for a
short target - and for any kind of UI or CLI.

The Guess class still allows different lengths, because the extra logic
required is so minimal, and nothing about that class itself prevents
different lengths from making sense.