Fix bug in _update_alphabet_states logic
Logic was assuming that an alphabet letter cannot go from UNUSED to
FOUND. However, this is incorrect, as it can happen within a single pass
of the function, if one instance of that letter is INCORRECT in the
guess and then a later instance is CORRECT in the guess. Reorganised
logic to make this more clear.