Add initial command line TUI using textual
There's a lot more to be done on this TUI before it matches the
functionality of the game implementation, but what's here so far is at
least fully working. At the same time, I would've perhaps preferred to
split this into separate features as individual commits, but it was all
getting hacked together at once while working through the textual docs.

I really some of textual, like the widgets making composing interfaces
simple. On the other hand, I think it might be a bit much for what I
need, and perhaps not too well suited to my situation. It doesn't seem
so well suited to having the program logic implemented in a class
outside of the textual definition (or at least the documentation doesn't
have any examples of doing this). The reactive attributes are cool but
it ended up quite clunky getting them to work the way I wanted. I still
have some problems with accessing the game during reursive composition
and modifying the message label after submitting a guess (the latter
should at least be solvable I think).

As previously the edges between WordleGame-specific functionality and
Game-generic functionality are a bit blurred, but the idea of how it
will work is growing now that I have an actual interface. The Guesses
Widget will probably need to be concrete-game specific, so that
different games can display the guesses differently. The same is
probably true for the alphabet, since the "status" will be different in
some way if there are e.g. multiple targets. Maybe there will be a way
to generalise multiple targets and therefore guesses, with Wordle as a
1-target case.

Another obvious missing part here is the configuration of the game
choice and options. I'd like that to be a separate interface screen, but
for now its just hard-coded. Tests are also missing, I need to read up
on how to implement them for textual.