Modify guess display and input styles to match
Much of this can be done by simple CSS, but adding spacing between
characters of the input requires a hack (see comment in code). To keep
the guess input generic I've added a new property to Game,
max_guess_word_length. The guess input may need to become game-specific
later anyway, but it seems like a reasonable property to have in the
interface.