Add numberle game and SingleWordleLikeBaseGame
Wordle and numberle end up looking very similar, but it doesn't make
much sense for numberle to inherit from wordle because of the word
dictionary stuff that's irrelevent for numberle - and besides, they're
more like peers. I considered some kind of composition approach to share
functionality without using inheritance, but in the end I think having a
common base class for games that have similar guess logic is simplest.
In the future I'll perhaps extend this base class to handle multiple
targets (which could just as well work for numberle as for wordle).

It makes most sense to test the functionality implemented in
SingleWordleLikeBaseGame just once, rather than duplicating for each
subclass. To make this work I've had to use a slightly complex system of
parametrization and dynamic fixture selection, but I think once you get
the hang of the pattern the end result is easy to follow. Some tests
have been slightly simplified to fit this pattern.

The tests make me think that the SingleWordleLikeBaseGame subclasses
could do with having their constructor parameters tidied up. Some could
use sensible default values, to make the object instantiation look
better aside from anything else. I especially think that the
"unspecified word length" option in Wordle is adding unnecessary
complexity for a feature that doesn't seem very useful (and it'll be
much worse when multiple targets are considered).

The TUI still needs to be updated, but the approach seems clearer now.