Import textual by module names
The code examples in the textual docs all import individual
functions/classes, so I tried it out like that. It's not like I've
specified a style guide in this repo, but in general I prefer improt
module names with few exceptions (e.g. typing) for the namespacing,
avoiding ambiguitiy and reducing how many names I need to remember the
source of. Textual doesn't seem worth making an exception for.