Generalise game-specific widget handling
Different subclasses of each widget kind can be defined where a widget
requires game-implementation-specific functionality (e.g. the `guesses`
attribute). Ideally these widgets would refer to protocols rather than
concrete Game subclasses (for duck typing, to avoid the widget
definitions being tied to whatever the class structure of the games is),
but python doesn't support types that are an intersection of a protocol
and an ABC yet. So for now each widget implemention has a type alias
referring to whatever game type(s) implement the functionality it
requires, at the risk of the widget accidentally using more
game-subclass-specific functionality than intended.

Actually using the widgets relies on a registry of the games and the
widgets they should use (the default would be empty widgets displaying
nothing). This registry is the main place (aside from type aliases
mentioned above) where the TUI is aware of specific games. I considered
having the games define their widgets themselves, but even if circular
imports could be avoided it didn't seem right for the games to be aware
of the TUI at all.

The main thing missing now is a way to configure game choice and
constructor (kw)args, so that games can be created in a more general
way. How that's implemented will inform whether the registry keys should
stay plain strings or become enums or something less error prone.

The tui tests probably need more attention to make them less
wordle-specific, and maybe even to test the widgets as their own units
(if that's even possible). Not sure if I'll actually spend much time on
that though.