Add instructions for testing workflows locally
Ending up with loads of commits while trying to fix a CI bug is a peeve
of mine, so a way to get feedback without committing is important (the
alternative would be squashing all those commits afterwards, which seems
messy). Github actions doesn't seem to have anything like jenkins'
replays, so running the pipeline locally first is the best approach I
found.