Add pre-commit lint CI job
The pre-commit action docs recommend using a newer pre-commit.ci github
"app" instead, but I found that: a) this didn't work easily with my mypy
setup (which relies on the environment outside of pre-commit, probably
not what they want me to do, but for good reason); and b) it meant
having the pre-commit CI output in a separate UI on the pre-commit site
itself, which I don't like. The recommended solution for a) is probably
pre-commit.ci lite, which just uses the old pre-commit action and adds
an extra step to commit back any auto-fixes. In fact, I don't really
care about the CI committing fixes, since I prefer to rely on local
linting and have the CI linting only fail in unusual circumstances. So
just failing the check is good enough. Which means that the security
concerns about the old pre-commit action aren't relevant, and it's the
simplest way to achieve what I want.

At first I thought about having the lint step in the same job as the
tests, which avoids repeating the setup steps. But I realised that this
would mean that if whichever comes first fails, the latter doesn't run.
It's not like I'm paying for those CI servers, so no harm in running
some steps twice for a better dev experience!