Remove pip-compile and requirements.txt
Since this is a library, there's no definite need for
frozen/pinned/locked dependencies. There are still good arguments for
it, and the idea of committing the lock even for libraries seems to be
gaining popularity. The main benifits would be making builds more
repeatable (it would be possible to see what dep versions the tests ran
on), preventing development halting if an upstream package releases a
bug or incompatible version, and making it easier to tell if test errors
are coming from upstream or internally. On the other hand, the primary
disadvantage is that tests no longer replicate the user experience of
installing the library in a fresh environment (in which they have no
lockfile of their own, so get the newest dependencies).

Testing from a lockfile and never updating the lockfile unless
absolutely necessary would go badly - the tests might go very long times
without replicating the user experience. Developers could make an effort
to manually update the lockfile, but that isn't likely with one dev on a
consistent dev environment. Other reasonable approaches might be to run
tests with both the locked and latest deps or the have some automation
to expire lockfiles/enforce updates. Both add some degree of complexity.
Running both reduces the benefits of the lockfile - builds are only
partly repeatable and development is still mostly halted by upstream
issues - because builds will still fail due to upstream issues. But
debugging is certainly much easier. Expiring the lockfile requires more
careful thought - does the build still pass, or does it require updating
the lockfile? Are lockfile updates automated or manual? If the build
passes, falling back to ignoring the lockfile, it would be tempting to
just leave it as-is and never update the lockfile - until a problem
occurs and the lockfile provides no real benefit.

For a larger project with more devs, some balanced approach that either
runs tests both with and without, or expires the lockfile regularly, is
probably better. In my case it doesn't seem worth the effort right now,
so I'll go with the simplest approach, which is still within the range
of reasonable recommendation. Maybe I'll regret later, or maybe I'll
come back and change it again. If I introduce tox or poetry I may find
that the balance of effort vs benefit changes.