Add package script definition for wordall command
When the package is installed, the TUI can be run with simply
`wordall`.

Also added some very basic tests in CI for this command existing and
running without exiting immediately. I considered some slightly more
involved ways to test that the command outputs something reasonable
(e.g. searching for the word "guess"), but due to the interactive nature
of the command this ended up becoming overly complex. I also felt that
there weren't any specific words I was sure would remain in the intial
TUI screen, especially if I later add a "select a game" menu, so I
didn't want to write a brittle test in CI. I also spent a bit of time
trying to make this kind of thing a unit test, but decided that too was
more effort than it was worth.