Use app_with_wordle_game for all tui tests
I inadvertently changed the default game type from numberle to wordle
which broke the tests in CI due to a dictionary file not existing. Some
of these tests could perhaps do with work to avoid this, but for now
using the fixture everywhere ensures that the existing get_game()
implementation is ignored and the dictionary is mocked.