Add SCOWL word list download script for package
SCOWL seems to be the most suitable word list to use for a default game
dictionary, and what I've been using for manual testing for a while.
The script skips downloading if the directory already exists mostly to
speed up dev and debugging - it isn't really important if the list is
outdated. I included the entire download because the irrelevant parts
relatively small, and having the original word lists organised in
separate files will allow a lot of flexibility, such as using common
words for the target but allowing guesses using any word.

The simplest way to add the SCOWL files to the package is to include
them within src/wordall, so that the relative path will be the same in
development and when installed. hatch-build-scripts is used to
conveniently run the download script within the build, although it
isn't very well documented.

Next there needs to be suitable configuration to use the SCOWL files by
default.