Move word dictionary loading to own class
The WordDictionaryLoader abstract class defines the interface for
loading a word dictionary, while the SimpleFileLoader defines a
specific way to achieve this. In the future there can be additional
loader implementations such as one to load from a combination of SCOWL
files.

This moves this responsibility away from the individual game, allows
easier re-use of word loading logic, and allows using different word
loading implementations in different situations. It also makes the
testing somewhat simpler, as we only need to worry about mocking file
access for testing that specific word loader - everywhere else we can
simply inject a mock word loader with a prepared output.