Add SCOWL word dictionary loader
The SCOWL loader can combine various selections of the word lists based
on their filenames. It is implemented as a subclass of a more generic
MultipleFileLoader.

A couple of other related refactorings also got tangled up here.

word_validator has been replaced with word_filter_function. In the
SCOWL case we don't have the option of creating a file with only valid
(alphabetical only) words, so all we can do is filter out the invalid
ones. The only disadvantage of this for other loaders is that mistakes
(like accidentally adding a symbol to the end of a line) may go
unnoticed.

The dictionary file loader exceptions have had to change to fit better
with the new classes. InvalidDictionaryWordError is no longer relevant
now that we are filtering. InvalidDictionaryFileError is now
NoWordsFoundError, handling the case where the final result has no
words. In a multiple file loader one of the files being empty is not
necessarily a problem, so checking only the output is more appropriate.

Creating fixtures for every possible combination of dictionary files
for MultipleFileLoader would be crazy, which led me to realise that
there was not much benefit to having these dictionary open() mocks as
fixtures rather than in a helper function (problem because of the tests
moving around since that pattern was first written). Using just a
helper function has the benefit of keeping the words themselves in the
test functions, increasing clarity of what's being tested. A similar
approach can be used for testing the new classes.