Add encoding option to word dictionary loaders
SCOWL uses iso8859-1 (latin-1) as its encoding, including some accents
which are incompatible with UTF-8. The SCOWL loader can just use
iso8859-1 all the time, but MultipleFileLoader still needs to support
encodings to enable it on the subclass, so might as well make it an
option on every loader class.